pyproject.toml
[build-system]
requires = ["setuptools>=62", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drl-trading-base"
version = "0.1.0"
description = "Base code for an algorithmic trading project prepared for Deep Reinforcement Learning (DRL)."
authors = [{ name="Your Name", email="you@example.com" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "streamlit>=1.36",
    "pandas>=2.0",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "ccxt>=4.3.0",
    "gymnasium>=0.29; python_version>='3.9'",
    "typing-extensions>=4.8",
    "tqdm>=4.65",
    "matplotlib>=3.7",
    "orjson>=3.9.7",
    # Optional
    "stable-baselines3>=2.3.2; python_version>='3.9'",
    "torch>=2.2; python_version>='3.9'",
    "polars>=1.0.0",
    "pyarrow>=15.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["tests"]
